refactor: CoreDataCloudMemo 싱글턴 방식 수정
- 기존에는 CoreDataCloudMemo타입이 싱글턴방식으로 구현했으나, 싱글턴을 굳이 쓰지않아도 될 것 같아 로직을 수정하였다.
- SplitViewController에서 CoreDataCloudMemo 인스턴스를 만들고 로직을 구현하도록 수정하였다.

+++ CloudNotes/CloudNotes/Utility/Protocol/CoreDatable.swift
@@ -11,7 +11,7 @@ protocol CoreDatable {
     var context: NSManagedObjectContext { get }
     
     func contextSave()
-    func deleteItem(object: NSManagedObject)
+    func deleteItem(object: NSManagedObject?)
     func updateItem(_ item: NSManagedObject, handler: (NSManagedObject) -> Void)
 }
 
@@ -29,7 +29,8 @@ extension CoreDatable {
         
     }
     
-    func deleteItem(object: NSManagedObject) {
+    func deleteItem(object: NSManagedObject?) {
+        guard let object = object else { return }
         context.delete(object)
     }
